fixed grandtotal error
