+++ zRecord/src/main/java/com/yona/estore/controller/HomeController.java
@@ -22,6 +22,10 @@ public class HomeController {
 	}
 
 	@RequestMapping("/login")
+//	@RequestParam=>this method parameter should be bound to a web request parameter
+//	"required = false" ==>Default is true, leading to an exception thrown in case of the parameter missing in the request. 
+//	Switch this to false if you prefer a null in case of the parameter missing.
+	
 	public String login(@RequestParam(value = "error", required = false) String error,
 			@RequestParam(value = "logout", required = false) String logout, Model model) {
 		if (error != null) {
