+++ zRecord/src/main/java/com/yona/estore/dao/CartDao.java
@@ -6,5 +6,5 @@ public interface CartDao {
 
 	Cart getCartById(int cartId);
 
-	void updateCart(Cart cart);
+	void update(Cart cart);
 }
