+++ zRecord/src/main/java/com/yona/estore/service/CartService.java
@@ -4,8 +4,7 @@ import com.yona.estore.model.Cart;
 
 public interface CartService {
 
-	Cart getCartById(int cartId);
-	
-	void updateCart(Cart cart);
+    Cart getCartById (int cartId);
 
+    void update(Cart cart);
 }
