+++ zRecord/src/main/java/com/yona/estore/service/Impl/CartItemServiceImpl.java
@@ -14,25 +14,21 @@ public class CartItemServiceImpl implements CartItemService {
 	@Autowired
 	private CartItemDao cartItemDao;
 	
-	@Override
 	public void addCartItem(CartItem cartItem) {
 		cartItemDao.addCartItem(cartItem);
 
 	}
 
-	@Override
 	public void removeCartItem(CartItem cartItem) {
 		cartItemDao.removeCartItem(cartItem);
 
 	}
 
-	@Override
 	public void removeAllCartItems(Cart cart) {
 		cartItemDao.removeAllCartItems(cart);
 
 	}
 	
-	@Override
 	public CartItem getCartItemByProductId(int productId) {
 		return cartItemDao.getCartItemByProductId(productId);
 	}
