+++ zRecord/src/main/java/com/yona/estore/service/Impl/CartServiceImpl.java
@@ -9,18 +9,15 @@ import com.yona.estore.service.CartService;
 
 @Service
 public class CartServiceImpl implements CartService {
-	
 
-	@Autowired
-	private CartDao cartDao;
+    @Autowired
+    private CartDao cartDao;
 
-	public Cart getCartById(int cartId) {
+    public Cart getCartById(int cartId) {
+        return cartDao.getCartById(cartId);
+    }
 
-		return cartDao.getCartById(cartId);
-	}
-
-	public void updateCart(Cart cart) {
-		cartDao.updateCart(cart);
-	}
-
+    public void update(Cart cart) {
+        cartDao.update(cart);
+    }
 }
