+++ zRecord/src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -28,6 +28,6 @@
     </bean>
     
     <mvc:resources mapping="/resources/**" location="/WEB-INF/resources/" />
-    
+    <!--  -->
     <tx:annotation-driven />
 </beans>
