+++ zRecord/src/main/webapp/WEB-INF/views/addProduct.jsp
@@ -10,23 +10,26 @@
 			<p class="lead">Fill the below information to add a product:</p>
 		</div>
 
-		<form:form
-			action="${pageContext.request.contextPath}/admin/product/addProduct"
-			method="post" commandName="product" enctype="multipart/form-data">
+		
+		<form:form	action="${pageContext.request.contextPath}/admin/product/addProduct"
+					method="post" commandName="product" 
+					enctype="multipart/form-data">
 			<div class="form-group">
 				<label for="name">Name</label>
 				<form:errors path="productName" cssStyle="color: #ff0000;" />
 				<form:input path="productName" id="name" class="form-Control" />
 			</div>
-
 			<div class="form-group">
-				<label for="category">Category</label> <label
-					class="checkbox-inline"> <form:radiobutton
-						path="productCategory" id="category" value="instrument" />Instrument
-				</label> <label class="checkbox-inline"><form:radiobutton
-						path="productCategory" id="category" value="record" />Record</label> <label
-					class="checkbox-inline"><form:radiobutton
-						path="productCategory" id="category" value="accessory" />Accessory</label>
+				<label for="category">Category</label> 
+				<label  class="checkbox-inline"> 
+					<form:radiobutton path="productCategory" id="category" value="instrument" />Instrument
+				</label> 
+				<label class="checkbox-inline">
+					<form:radiobutton path="productCategory" id="category" value="record" />Record
+				</label> 
+				<label class="checkbox-inline">
+					<form:radiobutton path="productCategory" id="category" value="accessory" />Accessory
+				</label>
 			</div>
 
 			<div class="form-group">
